Refactored cade, added inserting db test

+++ src/main/scala/model/Model.scala
@@ -4,8 +4,8 @@ package model
   * Created by rdsel on 31/10/2016.
   */
 
-case class Item(id: Long, date: String, name: String, price: Double, category: Int, userAssigned: Int)
-case class User(id: Int, fullName: String, branch: Int)
-case class Branch(id: Int, branch: String)
-case class Category(id: Int, category: String)
+case class Item(id: Long, date: String, name: String, price: Double, category: Long, userAssigned: Long)
+case class User(id: Long, fullName: String, branch: Long)
+case class Branch(id: Long, branch: String)
+case class Category(id: Long, category: String)
